[refactor] introduce debug classes under new structure
