+++ docs/extensions.rst
@@ -20,16 +20,16 @@ MSSQL       Untested  mssql:   native, PDO
 Other Interesting Drivers
 -------------------------
 
-===================== ========= ======== ============
-Class                 Support   PDO      Dependency
-===================== ========= ======== ============
-Connection_Dumper     Full      dumper:  native, Proxy
-Connection_Counter    Full      counter: native, Proxy
-===================== ========= ======== ============
+=========================== ========= ========== ============
+Class                       Support   PDO        Dependency
+=========================== ========= ========== ============
+Debug\Stopwatch\Driver      Full      stopwatch: native, Proxy
+Debug\Profiler\Driver       Full      profile:   native, Proxy
+=========================== ========= ========== ============
 
 
 3rd party vendor support
--------------------------
+------------------------
 
 ===================== ========= =========  ============================
 Class                 Support   PDO        Dependency
